requirements.txt
flask
flask-bcrypt
flask_login
flask_migrate
flask_sqlalchemy
flask_wtf
email_validator
Pillow